factored out message feed into a component
