+++ src/com/index.js
@@ -166,5 +166,6 @@ exports.message = require('./message')
 exports.messageThread = require('./message-thread')
 exports.messageSummary = require('./message-summary')
 exports.messagePreview = require('./message-preview')
+exports.messageFeed = require('./message-feed')
 exports.peers = require('./peers')
 exports.postForm = require('./post-form')
